update: MySQL 기반 채팅 기능 구현 v1 패키지로 이동
